requirements.txt
opencv-python>=4.13.0.92
numpy>=2.4.2
pyserial>=3.5
PyYAML>=6.0.3

pytest>=9.0.2
pytest-cov>=7.0.0
pytest-xdist>=3.8.0